Move profile pref registrations to brave_wallet_prefs
