+++ chromium_src/ios/chrome/browser/prefs/browser_prefs.mm
@@ -4,13 +4,13 @@
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "brave/components/brave_sync/brave_sync_prefs.h"
-#include "brave/components/brave_wallet/browser/keyring_controller.h"
+#include "brave/components/brave_wallet/browser/brave_wallet_prefs.h"
 #include "components/pref_registry/pref_registry_syncable.h"
 
 void BraveRegisterBrowserStatePrefs(
     user_prefs::PrefRegistrySyncable* registry) {
   brave_sync::Prefs::RegisterProfilePrefs(registry);
-  brave_wallet::KeyringController::RegisterProfilePrefs(registry);
+  brave_wallet::RegisterProfilePrefs(registry);
 }
 
 #define BRAVE_REGISTER_BROWSER_STATE_PREFS BraveRegisterBrowserStatePrefs(registry);
