+++ components/brave_wallet/browser/BUILD.gn
@@ -16,6 +16,8 @@ static_library("browser") {
     "asset_ratio_response_parser.h",
     "brave_wallet_constants.cc",
     "brave_wallet_constants.h",
+    "brave_wallet_prefs.cc",
+    "brave_wallet_prefs.h",
     "brave_wallet_provider_delegate.h",
     "brave_wallet_provider_impl.cc",
     "brave_wallet_provider_impl.h",
